requirements.txt
pandas~=1.5.1
python-dotenv~=0.21.0
selenium~=4.5.0
beautifulsoup4~=4.10.0
packaging==21.3
webdriver-manager==3.8.4